+++ lib/db/src/schema/index.ts
@@ -19,3 +19,4 @@
 
 export * from "./initiatives";
 export * from "./validations";
+export * from "./calculations";